Move another server-only file to its corresponding package
