+++ samples/validation/validation-shared/src/main/java/com/google/gwt/sample/validation/server/ServerGroup.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.google.gwt.sample.validation.shared;
+package com.google.gwt.sample.validation.server;
 
 /**
  * Server Validation Group.
